refactor: organizing types
This commit simply reorganizes the existing types into their own files.
Other than moving types to their own file, no other code was updated in
this specific commit.

This work is being done as there are plans to introduce more methods
defined on each type, e.g.: sorting methods, so that we are able to
sort an entire Kong config payload (which is useful for deterministic
comparisons, like in tests). Without this change, the `types.go`
file would get a bit out of control.
